Add files via upload
In cancer studies, typical research questions include:
(1) What is the impact of certain clinical characteristics on patient’s survival? For 
example, is there any difference between the group of people who has higher blood 
sugar and those who don’t?
(2) What is the probability that an individual survives a specific period (years, 
months, days)? For example, given a set of cancer patients, we will be able to tell that 
if 300(random number) days after the diagnosis of cancer has been passed, then the
probability of that person being alive at that time will be 0.7 (random number).
(3) Are there differences in survival between groups of patients? For example, let’s 
say there are 2 groups of people diagnosed with cancer. Those 2 groups were given 2 
different kinds of treatments. Now our goal here will be to find out if there is a 
significant difference between the survival time for those 2 different groups based on 
the treatment they were given.


Conclusion
From our analysis we find that as the survival days increases the 
probability of survival decreases while the probability of a person die 
increases. We also conclude that the value of hazard function increases as 
the survival time increases. 
We also conclude that the survival probability of female is less than the 
survival probability of male , thus doctor should focus on factor that 
affects the survival of male and give medicines accordingly.
Log rank test confirms that gender of a person affects the survival days of 
the person.
From cox regression we conclude that among the various factors gender 
and Ph.ECOG value affect most on the survival of patients while age of 
patient does not affect the survival of the patients. Thus , medicines must 
be prescribed keeping this factor in mind. Also, person having higher Ph. 
ECOG value has lower survival probability. Thus, medicines must be 
prescribed as such that it lowers the Ph. ECOG value.
Also, as the survival days increase the median survival time decreases